Added Lexical Analyzer
Added a lexical analyzer that uses PLY. Can create new objects of the
available data types(Edification, Schoo, Street, etc...), return the
type of an object, create new objects with no parameters and a partial
implementation of a for loop.
The file was originally from a calculator implementaion so it has other
useless methods that need to be removed.